fix bug

+++ Controllers/followConsole.js
@@ -16,18 +16,21 @@ res.status(200).json(response);
 const deleteFriend=async(req,res)=>{
     const { firstId, secondId } = req.body;
     try {
-        const delfriend = await followModel.findOneAndDelete({
+        await followModel.findOneAndDelete({
             members: { $all: [firstId, secondId] }
         })
         const delchat = await chatModel.findOneAndDelete({
             members: { $all: [firstId, secondId] }
         })
+        if(delchat){
+            const chatId = delchat._id;
+            await msgModel.deleteMany({ chatId: chatId })
+        }
+       
         const friend=await followModel.find({
             members:{$all:[firstId]}
         })
-        
-        const chatId = delchat._id
-        await msgModel.deleteMany({ chatId: chatId })
+       
        
         res.status(200).json(friend)
     }catch(err){
